+++ docs/changelog.rst
@@ -21,6 +21,7 @@ Currently in development.
 - Allow searching for tags in dropdown object selection fields
 - Renamed projects to project groups and groups to basic groups to avoid ambiguity
 - Allow disabling of subprojects / child project groups
+- Allow giving basic or project groups initial permissions
 
 Version 0.16.1
 --------------